feat: add files as a separate loading step
